// src/app/interfaces/reunion.ts
export interface Reunion {
  id_reunion: number;
  id_agenda: number;
  hora_inicio_agenda: string;
  hora_fin_agendad: string;
  fecha_agendada: string;
  hora_inicio_reunion: string;
  hora_fin_reunion: string;
  url_archivo_reunion: string;
  temas: string;
  acuerdos: string;
  observacion: string;
  estado: string;
  tipo_reunion: string;
  url_reunion: string;
  id_actividad_inscripcion: string;
}